Converted to support .Net standard & .Net 4.7.1
NOTE There are no tests for the task
