+++ Frends.Community.DownloadFile.Tests/Frends.Community.DownloadFile.Tests.cs
@@ -1,9 +1,10 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+using NUnit.Framework;
+using System;
 
 namespace Frends.Community.DownloadFile.Tests
 {
-    [TestClass()]
-    public class DownloadTests
+    [TestFixture]
+    class TestClass
     {
         // No test here because build server doesn't allow outgoing connections. 
     }
