--- Frends.Community.DownloadFile.Tests/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="SimpleImpersonation" version="3.0.0" targetFramework="net45" />
-</packages>
