+++ Frends.Community.DownloadFile.sln
@@ -1,11 +1,17 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.24720.0
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29613.14
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Frends.Community.DownloadFile", "Frends.Community.DownloadFile\Frends.Community.DownloadFile.csproj", "{C5C08D36-2991-4FBF-99BB-55C442AB2C99}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Frends.Community.DownloadFile", "Frends.Community.DownloadFile\Frends.Community.DownloadFile.csproj", "{35C305C0-8108-4A98-BB1D-AFE5C926239E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Frends.Community.DownloadFile.Tests", "Frends.Community.DownloadFile.Tests\Frends.Community.DownloadFile.Tests.csproj", "{B5F29694-511C-47F2-8253-DE08F0BACA8F}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Frends.Community.DownloadFile.Tests", "Frends.Community.DownloadFile.Tests\Frends.Community.DownloadFile.Tests.csproj", "{8CA92187-8E4F-4414-803B-EC899479022E}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{78F7F22E-6E20-4BCE-8362-0C558568B729}"
+	ProjectSection(SolutionItems) = preProject
+		.gitignore = .gitignore
+		README.md = README.md
+	EndProjectSection
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -13,16 +19,19 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{C5C08D36-2991-4FBF-99BB-55C442AB2C99}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C5C08D36-2991-4FBF-99BB-55C442AB2C99}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C5C08D36-2991-4FBF-99BB-55C442AB2C99}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C5C08D36-2991-4FBF-99BB-55C442AB2C99}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B5F29694-511C-47F2-8253-DE08F0BACA8F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B5F29694-511C-47F2-8253-DE08F0BACA8F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B5F29694-511C-47F2-8253-DE08F0BACA8F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B5F29694-511C-47F2-8253-DE08F0BACA8F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{35C305C0-8108-4A98-BB1D-AFE5C926239E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{35C305C0-8108-4A98-BB1D-AFE5C926239E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{35C305C0-8108-4A98-BB1D-AFE5C926239E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{35C305C0-8108-4A98-BB1D-AFE5C926239E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8CA92187-8E4F-4414-803B-EC899479022E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8CA92187-8E4F-4414-803B-EC899479022E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8CA92187-8E4F-4414-803B-EC899479022E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8CA92187-8E4F-4414-803B-EC899479022E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {55BC6629-85C9-48D8-8CA2-B0046AF1AF4B}
 	EndGlobalSection
 EndGlobal
