--- Frends.Community.DownloadFile/Frends.Community.DownloadFile.nuspec
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<package xmlns="http://schemas.microsoft.com/packaging/2011/10/nuspec.xsd">
-  <metadata>
-    <id>Frends.Community.DownloadFile</id>
-    <version>1.0.9</version>
-    <title>Frends.Community.DownloadFile</title>
-    <authors>HiQ Finland</authors>
-    <owners />
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <description>FRENDS community task for downloading files.</description>
-    <summary />
-    <dependencies>
-      <dependency id="SimpleImpersonation" version="3.0.0" />
-    </dependencies>
-    <frameworkAssemblies>
-      <frameworkAssembly assemblyName="System" />
-      <frameworkAssembly assemblyName="System.Core" />
-      <frameworkAssembly assemblyName="System.Data" />
-      <frameworkAssembly assemblyName="System.Data.DataSetExtensions" />
-    </frameworkAssemblies>
-  </metadata>
-  <files>
-    <file src="bin\release\Frends.Community.DownloadFile.dll" target="lib\net450\Frends.Community.DownloadFile.dll" />
-    <file src="bin\release\Frends.Community.DownloadFile.xml" target="lib\net450\Frends.Community.DownloadFile.xml" />
-  </files>
-</package>
