+++ Frends.Community.DownloadFile/FrendsTaskMetadata.json
@@ -1,0 +1,7 @@
+﻿{
+	"Tasks": [
+		{
+			"TaskMethod": "Frends.Community.DownloadFile.DownloadFile.DownloadFile"
+		}
+	]
+}
